# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileFitnessGoal = typeof ProfileFitnessGoal[keyof typeof ProfileFitnessGoal];


export const ProfileFitnessGoal = {
  lose_weight: 'lose_weight',
  build_muscle: 'build_muscle',
  improve_endurance: 'improve_endurance',
  stay_active: 'stay_active',
} as const;

export interface Profile {
  id: number;
  clerkUserId: string;
  name: string;
  email: string;
  age: number;
  weightKg: number;
  heightCm: number;
  fitnessGoal: ProfileFitnessGoal;
  createdAt: string;
}

export type ProfileInputFitnessGoal = typeof ProfileInputFitnessGoal[keyof typeof ProfileInputFitnessGoal];


export const ProfileInputFitnessGoal = {
  lose_weight: 'lose_weight',
  build_muscle: 'build_muscle',
  improve_endurance: 'improve_endurance',
  stay_active: 'stay_active',
} as const;

export interface ProfileInput {
  name: string;
  email: string;
  age: number;
  weightKg: number;
  heightCm: number;
  fitnessGoal: ProfileInputFitnessGoal;
}

export type ProfileUpdateFitnessGoal = typeof ProfileUpdateFitnessGoal[keyof typeof ProfileUpdateFitnessGoal];


export const ProfileUpdateFitnessGoal = {
  lose_weight: 'lose_weight',
  build_muscle: 'build_muscle',
  improve_endurance: 'improve_endurance',
  stay_active: 'stay_active',
} as const;

export interface ProfileUpdate {
  name?: string;
  age?: number;
  weightKg?: number;
  heightCm?: number;
  fitnessGoal?: ProfileUpdateFitnessGoal;
}

export type WorkoutType = typeof WorkoutType[keyof typeof WorkoutType];


export const WorkoutType = {
  running: 'running',
  cycling: 'cycling',
  swimming: 'swimming',
  weightlifting: 'weightlifting',
  yoga: 'yoga',
  hiit: 'hiit',
  walking: 'walking',
  other: 'other',
} as const;

export interface Workout {
  id: number;
  userId: string;
  type: WorkoutType;
  durationMinutes: number;
  caloriesBurned: number;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt: string;
}

export type WorkoutInputType = typeof WorkoutInputType[keyof typeof WorkoutInputType];


export const WorkoutInputType = {
  running: 'running',
  cycling: 'cycling',
  swimming: 'swimming',
  weightlifting: 'weightlifting',
  yoga: 'yoga',
  hiit: 'hiit',
  walking: 'walking',
  other: 'other',
} as const;

export interface WorkoutInput {
  type: WorkoutInputType;
  durationMinutes: number;
  caloriesBurned: number;
  notes?: string;
  date: string;
}

export interface NutritionEntry {
  id: number;
  userId: string;
  foodName: string;
  calories: number;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  carbs?: number | null;
  /** @nullable */
  fat?: number | null;
  date: string;
  createdAt: string;
}

export interface NutritionInput {
  foodName: string;
  calories: number;
  protein?: number;
  carbs?: number;
  fat?: number;
  date: string;
}

export interface DashboardStats {
  totalWorkoutsThisWeek: number;
  totalCaloriesBurnedThisWeek: number;
  totalCaloriesConsumedToday: number;
  currentWeightKg: number;
  workoutStreak: number;
  goalProgress: number;
}

export interface DailyStats {
  date: string;
  caloriesBurned: number;
  caloriesConsumed: number;
  workoutMinutes: number;
}

export interface WeightEntry {
  id: number;
  userId: string;
  weightKg: number;
  date: string;
  createdAt: string;
}

export interface WeightInput {
  weightKg: number;
  date: string;
}

export type ListWorkoutsParams = {
limit?: number;
offset?: number;
};

export type ListNutritionParams = {
date?: string;
};