# lib/api-zod/src/generated/types/dailyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyStats {
  date: string;
  caloriesBurned: number;
  caloriesConsumed: number;
  workoutMinutes: number;
}