# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalWorkoutsThisWeek: number;
  totalCaloriesBurnedThisWeek: number;
  totalCaloriesConsumedToday: number;
  currentWeightKg: number;
  workoutStreak: number;
  goalProgress: number;
}