# lib/api-zod/src/generated/types/nutritionEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface NutritionEntry {
  id: number;
  userId: string;
  foodName: string;
  calories: number;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  carbs?: number | null;
  /** @nullable */
  fat?: number | null;
  date: Date;
  createdAt: Date;
}