# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileFitnessGoal } from './profileFitnessGoal';

export interface Profile {
  id: number;
  clerkUserId: string;
  name: string;
  email: string;
  age: number;
  weightKg: number;
  heightCm: number;
  fitnessGoal: ProfileFitnessGoal;
  createdAt: Date;
}