# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileUpdateFitnessGoal } from './profileUpdateFitnessGoal';

export interface ProfileUpdate {
  name?: string;
  age?: number;
  weightKg?: number;
  heightCm?: number;
  fitnessGoal?: ProfileUpdateFitnessGoal;
}