# lib/api-zod/src/generated/types/weightEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface WeightEntry {
  id: number;
  userId: string;
  weightKg: number;
  date: Date;
  createdAt: Date;
}