# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutType } from './workoutType';

export interface Workout {
  id: number;
  userId: string;
  type: WorkoutType;
  durationMinutes: number;
  caloriesBurned: number;
  /** @nullable */
  notes?: string | null;
  date: Date;
  createdAt: Date;
}