# lib/api-zod/src/generated/types/workoutInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutInputType } from './workoutInputType';

export interface WorkoutInput {
  type: WorkoutInputType;
  durationMinutes: number;
  caloriesBurned: number;
  notes?: string;
  date: Date;
}